moves Groups dir to componets folder and renames it to GroupManagement
